INCOCHAIN - Tracking and Tracing Example
Be it document (bill of lading example) or a pallet of air freight, or LCL or Full Container of ocean cargo, or freight in the trucking industry, OR ANY DOCUMENT, here is an example of bill of lading, but it can be used by Account, sub account, by shipment number, or by any document. Not only is INCOCHAIN offering smart contracts for world trade, but also tracking and tracing capabilities for completely paperless business be it import or export, or even domestic leg of transportation industry.